Caches the result of `AutoCompleteSystem._project_users`
The method is called once every page load and makes several sql queries
for each project member which can be problematic in large projects.
While I think the previous commit should make the performance of this
method a lot better it still won't scale well in projects with many
members.

The method has an argument `all` that is never used, should it ever be
used the cache is bypassed altogether.  The default timeout of the cache
has been set to three minutes, it shouldn't be a big deal if there is a
name too many or too few in the autocomplete inputs for a while.

The cache is cleared when any of the methods of the interface
IGroupMembershipChangeListener is called.

The cached value is mutable.  In `AutoCompleteSystem` it is translated
to JSON and otherwise not used.  The method is also called from the
taskboard plugin, the return value is not modified there either.

Fixes #3723

git-svn-id: file:///mnt/ramdisk/d4-svn/trunk@21176 484df14f-75d2-4b3f-8ba0-490e05069a01